Changed stuffs

+++ src/com/gmail/amaarquadri/beast/connectr/server/logic/LocationData.java
@@ -31,7 +31,7 @@ public class LocationData implements Serializable {
         return longitude;
     }
 
-    public double getLastUpdateUnixTime() {
+    public long getLastUpdateUnixTime() {
         return lastUpdateUnixTime;
     }
 }
